refactor: isolate validation logic into Form Request classes
- Create LoginRequest, StoreCheckoutRequest, StoreJournalRequest

- Refactor AuthController, CheckoutController, JournalController to use Form Requests

- Add architecture_decisions.md for project memory

- Fix TestCase.php: use RefreshDatabase trait for proper SQLite testing

- Fix CheckoutTest assertion bug (expected 2 items but only 1 sent)

- All 12 tests passing with 58 assertions

+++ database/migrations/2026_09_13_000002_create_sale_items_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
             $table->foreignId('sale_id')->constrained('sales')->cascadeOnDelete();
             $table->foreignId('product_id')->constrained('products');
             $table->integer('quantity');
-            $table->unsignedBigInteger('price');
-            $table->unsignedBigInteger('subtotal');
+            $table->integer('price');
+            $table->integer('subtotal');
             $table->timestamps();
 
             $table->index(['sale_id', 'product_id']);
